docs: update with-next-translate example using appdir
