+++ examples/with-next-translate/i18n.json
@@ -3,6 +3,6 @@
   "defaultLocale": "en",
   "pages": {
     "*": ["common"],
-    "/": ["home"]
+    "/[lang]": ["home"]
   }
 }
