+++ examples/with-next-translate/next.config.js
@@ -1,3 +1,3 @@
-const nextTranslate = require('next-translate')
+const nextTranslate = require('next-translate-plugin')
 
 module.exports = nextTranslate({})
