+++ examples/with-next-translate/package.json
@@ -6,9 +6,12 @@
     "start": "next start"
   },
   "dependencies": {
-    "next": "10.0.3",
-    "next-translate": "1.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "next": "13.4.7",
+    "next-translate": "2.4.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "next-translate-plugin": "2.4.0"
   }
 }
